Move routes into separate module to prepare lazy loading
