+++ app/assets/stylesheets/layout/work_packages/_table.sass
@@ -44,7 +44,8 @@
 // div style="clear:both;" (is pushed to overflow - of no relevance)
 //
 // This makes use of more specific rules overwriting less specific ones.
-// Per default, the height is always 100%
+// Per default, the height is always 100
+.openproject-base--ui-view,
 .work-packages-page--ui-view
   height: 100%
