+++ app/views/layouts/angular.html.erb
@@ -34,7 +34,7 @@ See docs/COPYRIGHT.rdoc for more details.
 <% end -%>
 
 <%= content_for :content_body do %>
-  <work-packages-base></work-packages-base>
+  <openproject-base></openproject-base>
 <% end -%>
 
 <%= render template: 'layouts/base', locals: local_assigns.merge({ no_action_menu: true }) %>
