+++ frontend/src/app/angular4-modules.ts
@@ -34,7 +34,6 @@ import {OPContextMenuService} from 'core-components/op-context-menu/op-context-m
 import {OpModalService} from 'core-components/op-modals/op-modal.service';
 import {CurrentProjectService} from 'core-components/projects/current-project.service';
 import {ProjectCacheService} from 'core-components/projects/project-cache.service';
-import {FirstRouteService} from 'core-components/routing/first-route-service';
 import {States} from 'core-components/states.service';
 import {ExpandableSearchComponent} from 'core-components/expandable-search/expandable-search.component';
 import {PaginationService} from 'core-components/table-pagination/pagination-service';
@@ -52,8 +51,6 @@ import {OpenprojectFieldsModule} from "core-app/modules/fields/openproject-field
 import {OpenprojectCommonModule} from "core-app/modules/common/openproject-common.module";
 import {CommentService} from "core-components/wp-activity/comment-service";
 import {OpDragScrollDirective} from "core-app/modules/common/ui/op-drag-scroll.directive";
-import {UIRouterModule} from "@uirouter/angular";
-import {initializeUiRouterConfiguration} from "core-components/routing/ui-router.config";
 import {OpenprojectPluginsModule} from "core-app/modules/plugins/openproject-plugins.module";
 import {ConfirmFormSubmitController} from "core-components/modals/confirm-form-submit/confirm-form-submit.directive";
 import {ProjectMenuAutocompleteComponent} from "core-components/projects/project-menu-autocomplete/project-menu-autocomplete.component";
@@ -70,13 +67,15 @@ import {CurrentUserService} from 'core-components/user/current-user.service';
 import {OpenprojectWorkPackagesModule} from 'core-app/modules/work_packages/openproject-work-packages.module';
 import {OpenprojectAttachmentsModule} from 'core-app/modules/attachments/openproject-attachments.module';
 import {OpenprojectEditorModule} from 'core-app/modules/editor/openproject-editor.module';
+import {OpenprojectRouterModule} from "core-app/modules/router/openproject-router.module";
+import {OpenprojectCalendarModule} from "core-app/modules/calendar/openproject-calendar.module";
 
 @NgModule({
   imports: [
-    // UI router routes configuration
-    UIRouterModule.forRoot(),
     // Commons
     OpenprojectCommonModule,
+    // Router module
+    OpenprojectRouterModule,
     // Hal Module
     OpenprojectHalModule,
 
@@ -87,18 +86,16 @@ import {OpenprojectEditorModule} from 'core-app/modules/editor/openproject-edito
 
     OpenprojectAttachmentsModule,
     OpenprojectWorkPackagesModule,
+
+    // Dynamic calendars functionality
+    OpenprojectCalendarModule,
+
     // Plugin hooks and modules
     OpenprojectPluginsModule,
     // Linked plugins dynamically generated by bundler
     LinkedPluginsModule,
   ],
   providers: [
-    {
-      provide: APP_INITIALIZER,
-      useFactory: initializeUiRouterConfiguration,
-      deps: [Injector],
-      multi: true
-    },
     {
       provide: APP_INITIALIZER,
       useFactory: initializeServices,
@@ -114,7 +111,6 @@ import {OpenprojectEditorModule} from 'core-app/modules/editor/openproject-edito
     PaginationService,
     OpenProjectFileUploadService,
     CurrentProjectService,
-    FirstRouteService,
     // Split view
     CommentService,
     // Context menus
