+++ frontend/src/app/modules/calendar/wp-calendar/wp-calendar.component.ts
