+++ frontend/src/app/modules/calendar/wp-embedded-calendar/wp-embedded-calendar-entry.component.ts
