+++ frontend/src/app/modules/work_packages/routing/wp-full-view/wp-full-view.html
