+++ frontend/src/app/modules/work_packages/routing/wp-split-view/wp-split-view.html
